backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-pymongo==2.3.0
pymongo==4.7.3
python-dotenv==1.0.1
werkzeug==3.0.3
# NLP
spacy==3.7.5
# Resume parsing
pdfplumber==0.11.1
python-docx==1.1.2
# Production server
gunicorn==22.0.0
# Build compatibility
setuptools>=65.0
# AI features (Resume Enhancer + Career Roadmap)
groq>=1.0.0